requirements.txt
progressbar
pytorch
torchvision
pytoune
numpy
pandas
matplotlib